Backend/requirements.txt
Flask==3.1.1
Flask-PyMongo==2.3.0
Flask-CORS==5.0.1
Flask-Bcrypt==1.0.1
PyJWT==2.10.1
pymongo==4.12.1
python-dotenv==1.1.0
gunicorn==23.0.0
dnspython==2.7.0
marshmallow==3.26.1
